pyproject.toml
[tool.poetry]
name = "traffcap"
version = "0.1.0"
description = ""
authors = ["Chris Read <centurix@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
aiosqlite = "^0.17.0"
mypy = "^0.910"
yoyo-migrations = "^7.3.2"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
